# Fix skipping of range deleted keys in BlockBasedTableIterator