--- a/table/block_based_table_reader.h
+++ b/table/block_based_table_reader.h
@@ -508,7 +508,9 @@ class BlockBasedTableIterator : public InternalIterator {
         icomp_(icomp),
         range_del_agg_(range_del_agg),
         file_meta_(file_meta),
-        range_tombstone_(Slice(), Slice(), 0),
+        range_tombstone_start_(nullptr),
+        range_tombstone_end_(nullptr),
+        range_tombstone_seq_(0),
         index_iter_(index_iter),
         pinned_iters_mgr_(nullptr),
         block_iter_points_to_real_block_(false),
@@ -604,25 +606,8 @@ private:
   // the current file, allowing us to skip over a swath of deleted keys.
   void InitRangeTombstone(const Slice& target);
 
-  std::string tombstone_internal_start_key() const {
-    std::string internal_key;
-    AppendInternalKey(&internal_key, {
-        range_tombstone_.start_key_, range_tombstone_.seq_, kTypeValue});
-    return internal_key;
-  }
-
-  std::string tombstone_internal_end_key() const {
-    std::string internal_key;
-    // We specify kMaxSequenceNumber instead of the tombstone's
-    // sequence number because internal keys are ordered by descending
-    // sequence number. Using kMaxSequenceNumber ensures we'll seek to
-    // a version of the key that is more recent than the
-    // tombstone. Note that the tombstone end-key is exclusive, so the
-    // tombstone doesn't apply to the end-key in any case.
-    AppendInternalKey(&internal_key, {
-        range_tombstone_.end_key_, kMaxSequenceNumber, kTypeValue});
-    return internal_key;
-  }
+  std::string tombstone_internal_start_key() const;
+  std::string tombstone_internal_end_key() const;
 
   Slice user_key() const {
     assert(Valid());
@@ -638,9 +623,11 @@ private:
   // The range tombstone that covers the current key. Note that this is a
   // cooked value, not the raw tombstone as retrieved from
   // RangeDelAggregator::GetTombstone(). In particular, the start and end keys
-  // will be cleared to indicate the tombstone extends past the beginning or
+  // will be nullptr to indicate the tombstone extends past the beginning or
   // end of the sstable.
-  RangeTombstone range_tombstone_;
+  const Slice* range_tombstone_start_;
+  const Slice* range_tombstone_end_;
+  SequenceNumber range_tombstone_seq_;
   InternalIterator* index_iter_;
   PinnedIteratorsManager* pinned_iters_mgr_;
   BlockIter data_block_iter_;