TableCache::NewIterator was failing to pass the file metadata to
BlockBasedTableReader::NewIterator, so all of the fancy code in
BlockBasedTableIterator to skip ranges of deleted keys was not being
used.

Changed BlockBasedTableIterator to represent range tombstone endpoints
that extend to infinity using a nullptr rather than an empty key. An
empty key is a valid key and doing anything with a key other than using
it for comparison using the user comparator is invalid.

A tweak to the existing `BenchmarkRocksDBDeleteRangeIterate` shows the benefit from this change:

```
name                                                       old time/op  new time/op  delta
RocksDBDeleteRangeIterate/entries=10/deleted=9-8           6.11µs ± 2%  5.45µs ± 3%  -10.80%  (p=0.000 n=10+10)
RocksDBDeleteRangeIterate/entries=1000/deleted=999-8        132µs ± 3%     6µs ± 4%  -95.35%  (p=0.000 n=9+10)
RocksDBDeleteRangeIterate/entries=100000/deleted=99999-8   12.4ms ± 1%   0.0ms ± 3%  -99.95%  (p=0.000 n=9+9)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/rocksdb/11)
<!-- Reviewable:end -->